+++ src/pages/dashboard/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import '../App.css'
+import '../../App.css'
 
 import { Link } from '@reach/router';
